# web/openapi.yaml
openapi: 3.1.0
info:
  title: Owlette API
  version: 2.3.1
  description: |
    The Owlette API lets you programmatically manage your fleet of Windows machines,
    deploy software, monitor processes, execute tools, and integrate with external systems
    via webhooks.

    ## Authentication

    All API requests require authentication via an **API key**. You can create API keys
    from the Owlette dashboard under **Settings → API Keys**, or via the
    [Create API Key](#tag/API-Keys/operation/createApiKey) endpoint.

    Include your API key in every request using one of these methods:

    - **Header** (recommended): `x-api-key: owk_your_key_here`
    - **Query parameter**: `?api_key=owk_your_key_here`

    API keys are scoped to your user account and inherit your site access permissions.

    ## Rate Limiting

    All endpoints are rate-limited. When you exceed the limit, you'll receive a `429 Too Many Requests`
    response with a `Retry-After` header indicating how long to wait.

    Default limit: **300 requests/hour** per API key.

    Response headers on every request:
    - `X-RateLimit-Limit` — Maximum requests per window
    - `X-RateLimit-Remaining` — Requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Errors

    All errors return a JSON object with an `error` field:

    ```json
    { "error": "Description of what went wrong" }
    ```

    Common HTTP status codes:
    | Code | Meaning |
...

...

...
  license:
...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /api/roosts/{roostId}/versions/{versionRef}:
    x-stub: true
    parameters:
      - $ref: '#/components/parameters/RoostIdPath'
      - $ref: '#/components/parameters/VersionRefPath'
      - $ref: '#/components/parameters/RoostVersionHeader'
    get:
      operationId: getVersion
      tags: [versions]
      summary: fetch full version body
      description: |
        Returns the full oci version body.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/SiteIdQuery'
      responses:
        '200':
          description: Version detail with full layer list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionDetail' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
    patch:
      operationId: updateVersionDescription
      tags: [versions]
      summary: edit a published version's description
      description: |
        Edits the `description` field on an already-published version.
        The version's content (files, chunks) is immutable once
        published — any other body field is rejected with
        `version_content_immutable`.
      security:
        - apiKey: []
        - bearerApiKey: []
        - firebaseIdToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteId, description]
              properties:
                siteId:
                  type: string
                description:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: |
                    New description. ≤ 500 chars; longer payloads are
                    rejected with 400. Empty / whitespace-only strings
                    are stored as null.
      responses:
        '200':
          description: Updated version metadata.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VersionSummary' }
        '400':
          description: |
            Validation failed. Returns 400 with code
            `version_content_immutable` if the body contains any field
            other than `siteId` / `description`, or `version_ref_malformed`
            if the path `versionRef` has an invalid grammar.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401': { $ref: '#/components/responses/Problem401' }
        '403': { $ref: '#/components/responses/Problem403' }
        '404': { $ref: '#/components/responses/Problem404' }
        '429': { $ref: '#/components/responses/Problem429' }
        '500': { $ref: '#/components/responses/Problem500' }
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: x-api-key
      in: header
      description: |
        Legacy api key header — supported for backwards compatibility.
        New integrations should use `bearerApiKey` instead
        (`Authorization: Bearer owk_live_...`).
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: owk_live_* / owk_test_*
      description: |
        Preferred auth mechanism for the public roost api. Send the
        scoped api key as `Authorization: Bearer owk_live_<secret>`
        (production) or `Authorization: Bearer owk_test_<secret>`
        (sandbox). See [authentication.md](../docs/api/authentication.md).
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: |
        Firebase ID token issued to the authenticated user or agent device.
        Used by dashboard + agent requests to roost endpoints. First-party
        clients only — external integrations use `apiKey` /
        `bearerApiKey` instead.
  parameters:
    SiteIdQuery:
      name: siteId
      in: query
      required: true
      schema:
        type: string
      description: The site ID
...
    RoostIdPath:
      name: roostId
      in: path
      required: true
      schema:
        type: string
      description: The roost ID (deploy unit).
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
      description: |
        Client-generated idempotency key. If the server has seen the same
        key on a request with the same body within the replay window, it
        returns the original response instead of re-executing.
    RoostVersionHeader:
      name: Roost-Version
      in: header
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: '2026-04-22'
      description: |
        Selects a dated api version. Optional in wave 1; required once
        a second version ships. See [`GET /api/version`](#tag/utility/operation/getApiVersion).
...
    VersionRefPath:
      name: versionRef
      in: path
      required: true
      schema:
        type: string
      description: |
        Addresses a version on the roost. Accepts any of:
          - a stable content-addressed id: `vrs_<hex>`
          - a version number: `3`, `#3`, `v3`, or `V3`
          - a roost-relative alias: `current`, `previous`, `first`
        Malformed refs return 400 `version_ref_malformed`;
        unmatched refs return 404 `version_not_found`.
...
  headers:
    RateLimitLimit:
      description: Window limit for the rate-limit bucket that fired for this request.
      schema:
        type: integer
    RateLimitRemaining:
      description: Calls remaining in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: Seconds until the current window fully resets.
      schema:
        type: integer
    RetryAfter:
      description: Seconds the client should wait before retrying (429 only).
      schema:
        type: integer
    RoostRateLimitedReason:
      description: |
        Which rate-limit bucket tripped (`global-rate`, `endpoint-rate`,
        `key-rate`, `site-concurrency`). Present only on 429.
      schema:
        type: string
        enum: [global-rate, endpoint-rate, key-rate, site-concurrency]
  responses:
...
    Problem401:
      description: Authentication required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Problem403:
      description: Forbidden (e.g. cross-site access).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem404:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    Problem429:
      description: Rate limited.
      headers:
        RateLimit-Limit:           { $ref: '#/components/headers/RateLimitLimit' }
        RateLimit-Remaining:       { $ref: '#/components/headers/RateLimitRemaining' }
        RateLimit-Reset:           { $ref: '#/components/headers/RateLimitReset' }
        Retry-After:               { $ref: '#/components/headers/RetryAfter' }
        Roost-Rate-Limited-Reason: { $ref: '#/components/headers/RoostRateLimitedReason' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
    Problem500:
      description: Internal server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  schemas:
...
    ProblemDetails:
      type: object
      description: |
        RFC 7807 problem+json envelope returned by all roost endpoints on
        error. Unlike the legacy `ErrorResponse`, this carries a stable
        machine-readable `type` URI, a correlation `requestId`, and
        optional field-level `errors`.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the problem category.
          example: https://owlette.app/problems/validation-failed
        title:
          type: string
          description: Short, stable human-readable summary.
          example: validation failed
        status:
          type: integer
          description: HTTP status code, mirrors the response status.
          example: 400
        detail:
          type: string
          description: Occurrence-specific explanation; safe to show to users.
        instance:
          type: string
          description: URI reference for this occurrence (often the route path).
        requestId:
          type: string
          description: Correlation ID for trace lookup; quote in support tickets.
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: |
            Field-level errors keyed by dotted JSON path (400/422 responses).
            Example: `{ "body.hashes": ["must be a non-empty array"] }`
    VersionSummary:
      type: object
      description: Entry in a roost's immutable version history.
      required: [versionId, versionNumber, createdAt, totalSize, totalFiles]
      properties:
        versionId:
          type: string
          description: Stable content-addressed id (`vrs_*`).
        versionNumber:
          type: integer
          description: Monotonic 1-indexed version number, per-roost.
        description:
          type: string
          nullable: true
          maxLength: 500
          description: |
            Optional plaintext "what changed?" note supplied at publish
            time or edited later via PATCH. Null when missing.
        versionUrl:
          type: string
          format: uri
          description: R2 URL of the version body (short-TTL signed).
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: UID of the user or agent that published this version.
        totalSize:
          type: integer
          format: int64
          description: Sum of all file sizes in bytes.
        totalFiles:
          type: integer
        parentVersionId:
          type: string
          nullable: true
...
    Version:
      description: |
        Full oci-style version body. Roost emits schemaVersion 2,
        mediaType `application/vnd.owlette.version.v1+json`. See
        [versions.md](../docs/api/versions.md).
      type: object
      required: [schemaVersion, mediaType, config, layers]
      properties:
        schemaVersion:
          type: integer
          enum: [2]
        mediaType:
          type: string
          enum: [application/vnd.owlette.version.v1+json]
        config:
          type: object
          required: [mediaType, digest, size]
          properties:
            mediaType: { type: string }
            digest:
              type: string
              pattern: '^sha256:[0-9a-f]{64}$'
            size:
              type: integer
              format: int64
              minimum: 0
        layers:
          type: array
          items:
            type: object
            required: [mediaType, path, digest, size]
            properties:
              mediaType: { type: string }
              path:      { type: string }
              digest:
                type: string
                pattern: '^sha256:[0-9a-f]{64}$'
              size:
                type: integer
                format: int64
                minimum: 0
    VersionDetail:
      description: Version metadata wrapper returned on GET.
      type: object
      required: [id, versionNumber, publishedAt, totalSize, totalFiles, isCurrent, version]
      properties:
        id:
          type: string
          description: Stable content-addressed id (`vrs_*`).
        versionNumber:
          type: integer
          description: Monotonic 1-indexed version number, per-roost.
        description:
          type: string
          nullable: true
          maxLength: 500
          description: |
            Optional plaintext "what changed?" note. Editable via PATCH.
        publishedAt:
          type: string
          format: date-time
        publishedBy:
          type: string
          description: User or api key id that published this version.
        totalSize:
          type: integer
          format: int64
        totalFiles:
          type: integer
        isCurrent:
          type: boolean
        version: { $ref: '#/components/schemas/Version' }
...